refactor: orders wave 6 slice C — near-cap application/suggestion splits (#4246)
Extract completed-work handlers, army-move picker enumeration, and merchant
purchase-land candidate scan into sibling libraries; clear Slice C shrink-only
grandfather entries from repo.orders_lib_source_file_size.

Refs #4246
