+++ packages/colonizethis_orders/lib/src/orders/orders_application_road_propagation.dart
@@ -1,0 +1,83 @@
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import 'package:colonizethis_world/colonizethis_world.dart';
+import 'orders_application_context.dart';
+
+/// Propagates road transport level to adjacent capital/port tiles.
+///
+/// Per SPEC/program/development-resolution.md: "build_road: set or upgrade
+/// transport level for tileKey ... and, if applicable, adjacent capital/port
+/// tiles per capital-and-connectivity.md."
+///
+/// When a road is built adjacent to the player's capital or a port, the
+/// transport level is also applied to those adjacent tiles.
+TileMapState propagateRoadToAdjacentCapitalOrPort({
+  required String tileKey,
+  required int nextLevel,
+  Player? player,
+  required WorldState worldState,
+  required Map<String, TileMapResult> tileMapByRegion,
+  required TileMapState tileState,
+}) {
+  var current = tileState;
+  if (player == null) return current;
+
+  final parsedTile = parseTileKeyCoordinates(tileKey);
+  if (parsedTile == null) return current;
+  final targetRegionId = parsedTile.regionId;
+  final targetX = parsedTile.x;
+  final targetY = parsedTile.y;
+
+  // Get player's capital tile key
+  final capitalTileKey = player.capitalTile?.toTileKey();
+
+  // Get all port tile keys
+  final portTileKeys = worldState.portsByProvinceSeaboard.values.toSet();
+
+  // Get the tile map for the region to check bounds
+  final tileMap = tileMapByRegion[targetRegionId];
+  if (tileMap == null) return current;
+
+  // Check 4 neighbours (north, east, south, west)
+  final neighbours = [
+    (targetX, targetY - 1),
+    (targetX + 1, targetY),
+    (targetX, targetY + 1),
+    (targetX - 1, targetY),
+  ];
+
+  for (final (nx, ny) in neighbours) {
+    // Check bounds
+    if (nx < 0 || nx >= tileMap.width || ny < 0 || ny >= tileMap.height) {
+      continue;
+    }
+
+    // Get the cell (province) at this position
+    final cellId = tileMap.cell(nx, ny);
+
+    // Build the adjacent tile key
+    final adjacentTileKey = CapitalTile.tileKey(
+      targetRegionId,
+      '$targetRegionId|$cellId',
+      nx,
+      ny,
+    );
+
+    // Check if adjacent tile is the player's capital or a port
+    final isCapital = adjacentTileKey == capitalTileKey;
+    final isPort = portTileKeys.contains(adjacentTileKey);
+
+    if (isCapital || isPort) {
+      ordersApplicationLog.d(
+        'build_road propagating level $nextLevel to adjacent ${isCapital ? "capital" : "port"} tile $adjacentTileKey',
+      );
+      final roadLevel = current.roadLevel(adjacentTileKey);
+      // Only upgrade, never downgrade.
+      if (nextLevel > roadLevel) {
+        current = current.setRoadLevel(adjacentTileKey, nextLevel);
+      }
+    }
+  }
+  return current;
+}
